+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-10T04:39:16.593Z",
+    "2026-03-10T09:42:36.666Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {